feat: implemented controllers and modified handlers

+++ Movies.APP/Domain/Movie.cs
@@ -11,7 +11,7 @@ public class Movie : Entity
     
     public DateTime? ReleaseDate { get; set; }
     
-    public decimal TotalRevenue { get; set; }
+    public double TotalRevenue { get; set; }
     
     public int DirectorId { get; set; }
