+++ Movies.APP/Features/Movies/MovieCreateHandler.cs
@@ -14,7 +14,7 @@ public class MovieCreateRequest : Request, IRequest<CommandResponse>
 
     public DateTime? ReleaseDate { get; set; }
 
-    public decimal TotalRevenue { get; set; }
+    public double TotalRevenue { get; set; }
 
     public int DirectorId { get; set; }
